+++ CHANGELOG.md
@@ -269,11 +269,11 @@
 - Reports in review can be moved into the awaiting changes state   
 - Transactions & Planned Disbursements cannot be edited if they are associated with an approved Report
 - BEIS users can move a Report into the approved state
-
 - `Call open date` and `Call close date` added to the create activity form, for levels C and D.
   This field is mandatory for new activities, but optional for activities marked as `ingested: true`
 - `Create activity` buttons that were not changed on previous PR, are changed to
   `Add activity` now.
+- RODA Identifiers can be added to activities on creation and when updating
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-14...HEAD
 [release-14]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-13...release-14